# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conserje - API de Gestão Condominial
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MoradorStatus = (typeof MoradorStatus)[keyof typeof MoradorStatus];

export const MoradorStatus = {
  ativo: "ativo",
  inativo: "inativo",
} as const;

export interface Morador {
  id: number;
  nome: string;
  apartamento: string;
  bloco: string;
  telefone: string;
  status: MoradorStatus;
  created_at?: string;
}

export type EncomendaStatusValido =
  (typeof EncomendaStatusValido)[keyof typeof EncomendaStatusValido];

export const EncomendaStatusValido = {
  pendente: "pendente",
  notificado: "notificado",
  retirado: "retirado",
  extraviado: "extraviado",
} as const;

export interface Encomenda {
  id: number;
  codigo_rastreio: string;
  morador_id: number;
  morador_nome?: string;
  morador_apartamento?: string;
  morador_bloco?: string;
  status_valido: EncomendaStatusValido;
  data_recebimento: string;
  foto_url?: string | null;
  ocr_confianca?: number | null;
}

export type UpdateEncomendaStatusStatusValido =
  (typeof UpdateEncomendaStatusStatusValido)[keyof typeof UpdateEncomendaStatusStatusValido];

export const UpdateEncomendaStatusStatusValido = {
  pendente: "pendente",
  notificado: "notificado",
  retirado: "retirado",
  extraviado: "extraviado",
} as const;

export interface UpdateEncomendaStatus {
  status_valido: UpdateEncomendaStatusStatusValido;
}

export type LogInteracaoTipoMsg =
  (typeof LogInteracaoTipoMsg)[keyof typeof LogInteracaoTipoMsg];

export const LogInteracaoTipoMsg = {
  texto: "texto",
  imagem: "imagem",
} as const;

export type LogInteracaoPerfilEmocional =
  (typeof LogInteracaoPerfilEmocional)[keyof typeof LogInteracaoPerfilEmocional];

export const LogInteracaoPerfilEmocional = {
  neutro: "neutro",
  colaborativo: "colaborativo",
  estressado: "estressado",
  revoltado: "revoltado",
  "mal-intencionado": "mal-intencionado",
  empolgado: "empolgado",
  passivo: "passivo",
  frustrado: "frustrado",
  confuso: "confuso",
} as const;

export interface LogInteracao {
  id: number;
  morador_id: number;
  morador_nome?: string;
  morador_apartamento?: string;
  tipo_msg: LogInteracaoTipoMsg;
  acao: string;
  perfil_emocional: LogInteracaoPerfilEmocional;
  resultado: string;
  created_at?: string;
}

export interface PerfilEmocionalStat {
  perfil_emocional: string;
  count: number;
}

export interface AcoesStat {
  data: string;
  total: number;
  textos: number;
  imagens: number;
}

export type AlertaDefComNivelRisco =
  (typeof AlertaDefComNivelRisco)[keyof typeof AlertaDefComNivelRisco];

export const AlertaDefComNivelRisco = {
  baixo: "baixo",
  medio: "medio",
  alto: "alto",
  critico: "critico",
} as const;

export type AlertaDefComTipoAmeaca =
  (typeof AlertaDefComTipoAmeaca)[keyof typeof AlertaDefComTipoAmeaca];

export const AlertaDefComTipoAmeaca = {
  violencia: "violencia",
  criminalidade: "criminalidade",
  emergencia: "emergencia",
  intimidacao: "intimidacao",
} as const;

export interface AlertaDefCom {
  id: number;
  nivel_risco: AlertaDefComNivelRisco;
  tipo_ameaca: AlertaDefComTipoAmeaca;
  descricao: string;
  resolucao?: string | null;
  data_alerta: string;
  arquivado: boolean;
  autoridades_acionadas: boolean;
}

export interface UpdateAlerta {
  arquivado?: boolean;
  autoridades_acionadas?: boolean;
  resolucao?: string;
}

export type ListMoradoresParams = {
  apartamento?: string;
  bloco?: string;
  status?: string;
};

export type ListEncomendasParams = {
  apartamento?: string;
  status_valido?: string;
  data_inicio?: string;
  data_fim?: string;
  morador_id?: number;
};

export type ListLogsParams = {
  morador_id?: number;
  perfil_emocional?: string;
  data_inicio?: string;
  data_fim?: string;
};

export type ListAlertasParams = {
  nivel_risco?: string;
  tipo_ameaca?: string;
  resolvido?: boolean;
};