# lib/api-zod/src/generated/types/acoesStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conserje - API de Gestão Condominial
 * OpenAPI spec version: 0.1.0
 */

export interface AcoesStat {
  data: string;
  total: number;
  textos: number;
  imagens: number;
}