# lib/api-zod/src/generated/types/encomenda.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conserje - API de Gestão Condominial
 * OpenAPI spec version: 0.1.0
 */
import type { EncomendaStatusValido } from "./encomendaStatusValido";

export interface Encomenda {
  id: number;
  codigo_rastreio: string;
  morador_id: number;
  morador_nome?: string;
  morador_apartamento?: string;
  morador_bloco?: string;
  status_valido: EncomendaStatusValido;
  data_recebimento: Date;
  foto_url?: string | null;
  ocr_confianca?: number | null;
}