# lib/api-zod/src/generated/types/logInteracao.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conserje - API de Gestão Condominial
 * OpenAPI spec version: 0.1.0
 */
import type { LogInteracaoPerfilEmocional } from "./logInteracaoPerfilEmocional";
import type { LogInteracaoTipoMsg } from "./logInteracaoTipoMsg";

export interface LogInteracao {
  id: number;
  morador_id: number;
  morador_nome?: string;
  morador_apartamento?: string;
  tipo_msg: LogInteracaoTipoMsg;
  acao: string;
  perfil_emocional: LogInteracaoPerfilEmocional;
  resultado: string;
  created_at?: Date;
}