# lib/api-zod/src/generated/types/perfilEmocionalStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conserje - API de Gestão Condominial
 * OpenAPI spec version: 0.1.0
 */

export interface PerfilEmocionalStat {
  perfil_emocional: string;
  count: number;
}