# lib/api-zod/src/generated/types/updateEncomendaStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conserje - API de Gestão Condominial
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateEncomendaStatusStatusValido } from "./updateEncomendaStatusStatusValido";

export interface UpdateEncomendaStatus {
  status_valido: UpdateEncomendaStatusStatusValido;
}